move configDump to debug
