+++ pilot/pkg/proxy/envoy/v2/ads.go
@@ -22,23 +22,20 @@ import (
 	"sync"
 	"time"
 
-	adminapi "github.com/envoyproxy/go-control-plane/envoy/admin/v2alpha"
 	xdsapi "github.com/envoyproxy/go-control-plane/envoy/api/v2"
 	core "github.com/envoyproxy/go-control-plane/envoy/api/v2/core"
 	ads "github.com/envoyproxy/go-control-plane/envoy/service/discovery/v2"
-	"github.com/golang/protobuf/ptypes/any"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/peer"
 	"google.golang.org/grpc/status"
 
 	"istio.io/istio/pkg/config/schemas"
+	istiolog "istio.io/pkg/log"
 
 	"istio.io/istio/pilot/pkg/features"
-
 	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pilot/pkg/networking/util"
-	istiolog "istio.io/pkg/log"
 )
 
 var (
@@ -116,56 +113,6 @@ type XdsConnection struct {
 	// added will be true if at least one discovery request was received, and the connection
 	// is added to the map of active.
 	added bool
-}
-
-// configDump converts the connection internal state into an Envoy Admin API config dump proto
-// It is used in debugging to create a consistent object for comparison between Envoy and Pilot outputs
-func (s *DiscoveryServer) configDump(conn *XdsConnection) (*adminapi.ConfigDump, error) {
-	dynamicActiveClusters := []*adminapi.ClustersConfigDump_DynamicCluster{}
-	clusters := s.generateRawClusters(conn.modelNode, s.globalPushContext())
-
-	for _, cs := range clusters {
-		dynamicActiveClusters = append(dynamicActiveClusters, &adminapi.ClustersConfigDump_DynamicCluster{Cluster: cs})
-	}
-	clustersAny, err := util.MessageToAnyWithError(&adminapi.ClustersConfigDump{
-		VersionInfo:           versionInfo(),
-		DynamicActiveClusters: dynamicActiveClusters,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	dynamicActiveListeners := []*adminapi.ListenersConfigDump_DynamicListener{}
-	listeners := s.generateRawListeners(conn, s.globalPushContext())
-	for _, cs := range listeners {
-		dynamicActiveListeners = append(dynamicActiveListeners, &adminapi.ListenersConfigDump_DynamicListener{Listener: cs})
-	}
-	listenersAny, err := util.MessageToAnyWithError(&adminapi.ListenersConfigDump{
-		VersionInfo:            versionInfo(),
-		DynamicActiveListeners: dynamicActiveListeners,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	routes := s.generateRawRoutes(conn, s.globalPushContext())
-	routeConfigAny := util.MessageToAny(&adminapi.RoutesConfigDump{})
-	if len(routes) > 0 {
-		dynamicRouteConfig := []*adminapi.RoutesConfigDump_DynamicRouteConfig{}
-		for _, rs := range routes {
-			dynamicRouteConfig = append(dynamicRouteConfig, &adminapi.RoutesConfigDump_DynamicRouteConfig{RouteConfig: rs})
-		}
-		routeConfigAny, err = util.MessageToAnyWithError(&adminapi.RoutesConfigDump{DynamicRouteConfigs: dynamicRouteConfig})
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	bootstrapAny := util.MessageToAny(&adminapi.BootstrapConfigDump{})
-	// The config dump must have all configs with connections specified in
-	// https://www.envoyproxy.io/docs/envoy/latest/api-v2/admin/v2alpha/config_dump.proto
-	configDump := &adminapi.ConfigDump{Configs: []*any.Any{bootstrapAny, clustersAny, listenersAny, routeConfigAny}}
-	return configDump, nil
 }
 
 // XdsEvent represents a config or registry event that results in a push.
